Complete coverage
I added tests that provide complete coverage.  The coverage isn't 100%
in unmarshal.go, but that is only because the logic of the code makes
certain errors unreachable.  I prefer to leave the code that checks
for these errors in just in case a refactoring breaks the logic that
makes them unreachable.

The great thing is that no bugs were uncovered as test coverage increased.